Moved type tag to sidebar folder.
