+++ components/project-detail-page/sidebar/project-detail-type-tag.tsx
